[REFACTOR] external 모듈 분리 #82
