+++ umbba-external/src/main/java/sopt/org/umbba/external/client/auth/apple/AppleApiClient.java
@@ -1,8 +1,8 @@
-package sopt.org.umbba.domain.user.social.apple.feign;
+package sopt.org.umbba.external.client.auth.apple;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
-import sopt.org.umbba.domain.user.social.apple.response.ApplePublicKeys;
+import sopt.org.umbba.external.client.auth.apple.response.ApplePublicKeys;
 
 @FeignClient(name = "apple-public-verify-client", url = "https://appleid.apple.com/auth")
 public interface AppleApiClient {
