+++ umbba-external/src/main/java/sopt/org/umbba/external/client/auth/kakao/response/KakaoAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package sopt.org.umbba.domain.user.social.kakao.response;
+package sopt.org.umbba.external.client.auth.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
