+++ umbba-external/src/main/java/sopt/org/umbba/external/client/auth/kakao/response/KakaoUserResponse.java
@@ -1,4 +1,4 @@
-package sopt.org.umbba.domain.user.social.kakao.response;
+package sopt.org.umbba.external.client.auth.kakao.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,7 +9,10 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoAccount {
+public class KakaoUserResponse {
 
-    private KakaoUserProfile profile;
+    //받아올 땐 Long이지만, String으로 바꿔서 사용하기
+    private Long id;
+
+    private KakaoAccount kakaoAccount;
 }
